refactor(download): refactor download module
Linked to #33
