+++ public/app.js
@@ -2,6 +2,7 @@ import angular from 'angular';
 import SearchModule from './search/search';
 import PodcastsModule from './podcasts/podcasts';
 import ItemModule from './item/item';
+import DownloadModule from './download/download';
 
 let app = angular.module('podcastApp', [
     SearchModule.name,
